ParsedLine.value is now f32

+++ src/converters/celsius.rs
@@ -1,7 +1,7 @@
-pub fn to_fahrenheit(value: u32) -> f32 {
-    (value as f32) * (9.0 / 5.0) + 32.0
+pub fn to_fahrenheit(value: f32) -> f32 {
+    value * (9.0 / 5.0) + 32.0
 }
 
-pub fn to_kelvins(value: u32) -> f32 {
-    (value as f32) + 273.15
+pub fn to_kelvins(value: f32) -> f32 {
+    value + 273.15
 }
